pyproject.toml
[project]
name = "twenty-crm-agent"
version = "0.1.0"
description = "AI agent that queries Twenty CRM via GraphQL using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-community>=0.3.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
crm-cli = "src.cli.main:main"
crm-web = "src.web.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
